pyproject.toml
[project]
name = "cti-agent-sample"
version = "0.1.0"
description = "Template cyber threat intel agent: fetch, triage, enrich, report, detections."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.scripts]
cti-agent = "cti_agent.cli:app"

[tool.ruff]
line-length = 100

